LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-content-type-string-expected.txt

CONSOLE MESSAGE: line 23: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: line 24: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

  

PASS XMLHttpRequest: send() - Content-Type 
FAIL XMLHttpRequest: send() - Content-Type 1 assert_equals: expected "application/xml;charset=UTF-8" but got "NO"
FAIL XMLHttpRequest: send() - Content-Type 2 assert_equals: expected "text/html;charset=UTF-8" but got "NO"
